+++ apps/engine/lib/engine/analyzer/aliases.ex
@@ -42,10 +42,8 @@ defmodule Engine.Analyzer.Aliases do
 
       [prefix | suffix] ->
         case aliases do
-          %{^prefix => _} ->
-            current_module = resolve_alias(aliases, prefix, suffix)
-
-            Module.concat([current_module | suffix])
+          %{[^prefix] => _} ->
+            resolve_alias(aliases, prefix, suffix)
 
           _ ->
             Module.concat(module)
@@ -60,7 +58,7 @@ defmodule Engine.Analyzer.Aliases do
   defp resolve_alias(aliases, prefix, suffix) do
     current_module =
       aliases
-      |> Map.get(prefix)
+      |> Map.get(List.wrap(prefix))
       |> Alias.to_module()
 
     Module.concat([current_module | suffix])