+++ apps/engine/lib/engine/application.ex
@@ -16,6 +16,7 @@ defmodule Engine.Application do
           Engine.Dispatch,
           Engine.ModuleMappings,
           Engine.Build,
+          Engine.ModuleStore,
           Engine.Build.CaptureServer,
           Engine.Plugin.Runner.Supervisor,
           Engine.Plugin.Runner.Coordinator,