+++ apps/expert/lib/expert/port.ex
@@ -190,7 +190,10 @@ defmodule Expert.Port do
 
     path =
       if shell_available?(shell_env) do
-        path_env_at_directory(root_path, shell_env)
+        case path_env_at_directory(root_path, shell_env) do
+          {:ok, path} -> path
+          {:error, :timeout} -> filter_release_root_from_path()
+        end
       else
         filter_release_root_from_path()
       end
@@ -260,35 +263,69 @@ defmodule Expert.Port do
       {"PATH", System.get_env("PATH", @default_unix_path)}
     ]
 
-    shell_name = Path.basename(shell)
+    args = path_fetch_cmd_args(shell, directory)
 
-    args =
-      case shell_name do
-        "fish" ->
-          cmd =
-            "cd #{directory}; printf \"#{@path_marker}:%s:#{@path_marker}\" (string join ':' $PATH)"
+    maybe_cmd_output =
+      case cmd_with_timeout(shell, args, env, 1_000) do
+        {:ok, result} ->
+          {:ok, result}
 
-          ["-l", "-c", cmd]
-
-        "nu" ->
-          cmd =
-            "cd #{directory}; print $\"#{@path_marker}:($env.PATH | str join \":\"):#{@path_marker}\""
-
-          ["-l", "-c", cmd]
-
-        _ ->
-          cmd = "cd #{directory} && printf \"#{@path_marker}:%s:#{@path_marker}\" \"$PATH\""
-          ["-i", "-l", "-c", cmd]
+        {:error, :timeout} ->
+          if Enum.member?(args, "-i") do
+            # If the command contained the -i flag, try again without it.
+            # Some users have exec calls or blocking prompts in their .bashrc,
+            # so we would hang here without the timeout
+            args = Enum.reject(args, &(&1 == "-i"))
+            cmd_with_timeout(shell, args, env, 1_000)
+          else
+            {:error, :timeout}
+          end
       end
 
-    {output, exit_code} = System.cmd(shell, args, env: env)
+    case maybe_cmd_output do
+      {:ok, {output, exit_code}} ->
+        case Regex.run(~r/#{@path_marker}:(.*?):#{@path_marker}/s, output) do
+          [_, clean_path] when exit_code == 0 ->
+            {:ok, clean_path}
 
-    case Regex.run(~r/#{@path_marker}:(.*?):#{@path_marker}/s, output) do
-      [_, clean_path] when exit_code == 0 ->
-        clean_path
+          _ ->
+            {:ok, output |> String.trim() |> String.split("\n") |> List.last()}
+        end
+
+      {:error, :timeout} ->
+        {:error, :timeout}
+    end
+  end
+
+  defp path_fetch_cmd_args(shell, directory) do
+    case Path.basename(shell) do
+      "fish" ->
+        cmd =
+          "cd #{directory}; printf \"#{@path_marker}:%s:#{@path_marker}\" (string join ':' $PATH)"
+
+        ["-l", "-c", cmd]
+
+      "nu" ->
+        cmd =
+          "cd #{directory}; print $\"#{@path_marker}:($env.PATH | str join \":\"):#{@path_marker}\""
+
+        ["-l", "-c", cmd]
 
       _ ->
-        output |> String.trim() |> String.split("\n") |> List.last()
+        cmd = "cd #{directory} && printf \"#{@path_marker}:%s:#{@path_marker}\" \"$PATH\""
+        ["-i", "-l", "-c", cmd]
+    end
+  end
+
+  defp cmd_with_timeout(shell, args, env, timeout) do
+    task = Task.async(fn -> System.cmd(shell, args, env: env) end)
+
+    case Task.yield(task, timeout) || Task.shutdown(task) do
+      {:ok, result} ->
+        {:ok, result}
+
+      _ ->
+        {:error, :timeout}
     end
   end